+++ src/data/gameplay_events/arctap.c
@@ -5,7 +5,7 @@
 #include "rlgl.h"
 #include "raymath.h"
 
-void draw_arctap(MeshRenderable *arctap_r, ArcTap *arctap, float z_pos)
+void draw_arctap(MeshRenderable *arctap_r, ArcTap *arctap, double z_pos)
 {
   float fade_ratio = (z_pos - SKY_STOP_FADE) / (TAP_START_FADE - SKY_STOP_FADE);
   arctap_r->material.maps[MATERIAL_MAP_DIFFUSE].color = Fade(WHITE, Clamp(fade_ratio, 0.0f, 1.0f));
